feat: pages premium, optout, rdva
